Move the project_index_and_invoke functor into its own header
